+++ dlcdb/lending/admin.py
@@ -2,7 +2,12 @@ from django.contrib import admin
 from django.db import models
 from django.forms import Textarea
 
-from .models import LendingConfiguration
+from .models import LendingConfiguration, LendingConfigurationRegulation
+
+
+class RegulationInline(admin.TabularInline):
+    model = LendingConfigurationRegulation
+    extra = 1
 
 
 @admin.register(LendingConfiguration)
@@ -23,14 +28,19 @@ class LendingConfigurationAdmin(admin.ModelAdmin):
                 "fields": ("lending_preparation_checklist",),
             },
         ),
-        (
-            "Regulations",
-            {
-                "classes": ("collapse",),
-                "fields": ("mandatory_regulations",),
-            },
-        ),
+        # (
+        #     "Regulations",
+        #     {
+        #         "classes": ("collapse",),
+        #         "fields": ("mandatory_regulations",),
+        #     },
+        # ),
     )
+
+    inlines = [
+        RegulationInline,
+    ]
+    exclude = ["mandatory_regulations"]
 
     def has_add_permission(self, request):
         return not LendingConfiguration.objects.exists()